// src/types/Theme.ts
export interface ThemeSettings {
  colors: {
    primary: string;
    secondary: string;
    accent: string;
    text: string;
    background: string;
    lightBackground: string;
  };
  fonts: {
    heading: string;
    body: string;
  };
  layout: string;
  spacing: string;
  borderRadius: string;
  shadows: string;
}